move karvonnen out of app.js and require it into run.app.js when loading app to the watch
