+++ apps/run/metadata.json
@@ -28,6 +28,10 @@
     {
       "name": "run.settings.js",
       "url": "settings.js"
+    },
+    {
+      "name": "run_karvonnen",
+      "url": "run_karvonnen.js"
     }
   ],
   "data": [
